+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [Batch transactions support](https://github.com/multiversx/mx-sdk-dapp/pull/722)]
+## [Added Batch transactions support](https://github.com/multiversx/mx-sdk-dapp/pull/722)]
 
 ## [[v2.10.17]](https://github.com/multiversx/mx-sdk-dapp/pull/721)] - 2023-04-04