+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[v2.11.0]](https://github.com/multiversx/mx-sdk-dapp/pull/725)] - 2023-04-12
+
 - [Added Batch transactions support](https://github.com/multiversx/mx-sdk-dapp/pull/722)]
 - [Updated WalletConnect provider package](https://github.com/multiversx/mx-sdk-dapp/pull/726)]